Download, Decompress, and Import SoyBase GFF3 Annotation Files

Automatically connects to the authoritative SoyBase repository data collection, downloads the compressed Williams 82 genome annotation layer (.gff3.gz), and extracts it into a memory-optimized data frame ready for variant structural mapping.

get_soy_gff_from_soybase(dest_dir = "data", overwrite = FALSE)

Arguments

dest_dir

Character string pointing to the local directory where the downloaded GFF3 file should be stored. Default is "data".

overwrite

Logical. If TRUE, overwrites the file if it already exists locally in the target folder. Default is FALSE.

Value

A standard data.frame containing parsed genomic features (such as chromosomes, start/end positions, strands, and associated metadata attributes). This function return gff3 file only for soybean. Download it manually for your crop of interest

Examples

if (FALSE) { # \dontrun{
library(LDdecay)

# Downloads from SoyBase, decompresses, and builds your matrix cleanly
gff_table <- get_soy_gff_from_soybase(dest_dir = "data")

# Preview the results instantly!
head(gff_table)
} # }